pyproject.toml
[project]
name = "milo-data-platform"
version = "0.1.0"
description = "Data platform for Scandalicious consumer panel data product"
requires-python = ">=3.11"
dependencies = [
    # Snowflake
    "snowflake-connector-python[pandas]>=3.6.0",
    "snowflake-sqlalchemy>=1.6.0",

    # Database
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",

    # Data processing
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",

    # APIs
    "requests>=2.31.0",
    "httpx>=0.27.0",

    # Vector DB & embeddings
    "pinecone-client>=3.0.0",
    "sentence-transformers>=3.0.0",

    # Environment
    "python-dotenv>=1.0.0",

    # Exports
    "jinja2>=3.1.0",
    "weasyprint>=62.0",

    # Scheduling (optional — only if using Dagster)
    # "dagster>=1.7.0",
    # "dagster-snowflake>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]
dagster = [
    "dagster>=1.7.0",
    "dagster-snowflake>=0.23.0",
    "dagster-webserver>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py311"
